refactor(cli): extract common helpers and move business logic to core
  - Add extract_filename() and create_ref_service() helpers to common.rs
  - Move get_partition_stats() and PartitionStats from stats.rs to core/analysis
  - Move test_catalog_connectivity() from doctor.rs CLI to DoctorService core
  - Update branch.rs, tag.rs, vacuum.rs, repair.rs, import.rs, diff.rs
    to use centralized helpers
  - Fix clippy warnings: collapse else-if blocks in branch/tag,
    remove unnecessary cast in inspect.rs

  This refactoring reduces code duplication across CLI commands by
  centralizing common operations in shared modules and moving
  business logic from CLI thin-wrappers to core services.

+++ src/core/operations/inspect.rs
@@ -376,7 +376,7 @@ impl IcebergTableInspector {
         // ═══════════════════════════════════════════════════════════════════════
         // TABLE INFORMATION
         // ═══════════════════════════════════════════════════════════════════════
-        let format_version = version as i32;
+        let format_version = version;
         let location = metadata.location().to_string();
         let table_uuid = metadata.uuid().to_string();
         let current_snapshot_id = metadata.current_snapshot_id();
